+++ src/Examples/CoreExample/Controllers/AccountController.cs
@@ -6,7 +6,7 @@ namespace Example.Controllers
 {
     public class AccountController : Controller
     {
-        private string _appId = Environment.GetEnvironmentVariable("YOTI_APPLICATION_ID");
+        private readonly string _appId = Environment.GetEnvironmentVariable("YOTI_APPLICATION_ID");
         public static byte[] PhotoBytes { get; set; }
 
         // GET: Account/Connect?token